update new tests introduced by the merge
