+++ tests/end-to-end/README.md
@@ -77,15 +77,17 @@ These tests cover:
    the name of the new test. Example: `tests/end-to-end/mytest/template.json`.
 
     *The name of the folder must be a valid Rust identifier.*
-1. Save the CloudFormation template that will be the input for the test to a
-   file at `tests/end-to-end/mytest/template.json`.
-1. If the template you want to test relies on referencing resources that must
+2. Save the CloudFormation template that will be the input for the test to a
+   file at `tests/end-to-end/mytest/template.json`. This template must be in
+   JSON, not YAML format, because `cdk synth` creates CloudFormation templates
+   in JSON. It's easier to compare these if they are in the same format.
+3. If the template you want to test relies on referencing resources that must
    already exist, also add a template that creates those dependency resources at
    `tests/end-to-end/mytest/create_first.json`. If this file is present, the
    tests will create a CloudFormation stack with this template, before
    attempting to create a CloudFormation stack with the template called
    `template.json`.
-1. Add your test case to `tests/end-to-end.rs`, by editing the file manually.
+4. Add your test case to `tests/end-to-end.rs`, by editing the file manually.
 
     The file will contain a series of calls to the `test_case!` macro that will
     look something like this:
@@ -122,7 +124,7 @@ These tests cover:
 
     // Add new test cases here
     ```
-1. Make sure the input CloudFormation template is a valid CloudFormation
+5. Make sure the input CloudFormation template is a valid CloudFormation
    template.
     1. Set the `CREATE_CFN_STACK` environment variable to true.
         
@@ -140,7 +142,7 @@ These tests cover:
         You must use `unset` and not set the variable to 0 or false, because the
         logic used in the tests just checks for absence or presence of a value
         in this env variable.
-    1. Run your test.
+    2. Run your test.
         
         `cargo test --test end-to-end mytest -- --nocapture`
 
@@ -155,15 +157,23 @@ These tests cover:
         After getting past this part, the test will fail on later parts of the
         workflow. Follow the next steps to resolve.
 
-1. CDK app files
-
-    Now that we know the input CloudFormation template is valid, we need an app
-    file for each language
+6. CDK app files
+   
+   Now that we know the input CloudFormation template is valid, we need an app
+   file for each language.
      
-    set UPDATE_SNAPSHOTS=true run your test again
+   Set UPDATE_SNAPSHOTS=true run your test again. This will generate a default
+   app file that instantiates a CDK app with one stack. The stack is initialized
+   with no construct props, meaning default values will be used. See step 8 for
+   more details about app files.
 
-1. If needed, iterate until your test passes for all languages. You can run a
-single test in a specific language. For example:
+   ```sh
+   export UPDATE_SNAPSHOTS=true
+   cargo test --test end-to-end mytest
+   ```
+
+7. If needed, iterate until your test passes for all languages. You can run a
+   single test in a specific language. For example:
 
     ```sh
     cargo test --test end-to-end mytest::typescript
@@ -177,7 +187,7 @@ single test in a specific language. For example:
     test_case!(simple, "SimpleStack", &["golang"]);
     ```
     
-1. Now that your test is passing, determine if this test case should be testing
+8. Now that your test is passing, determine if this test case should be testing
    different combinations of Stack Construct props. If the input CloudFormation
    template had parameters, then `cdk-from-cfn` will generate a CDK stack with
    some custom Stack props. Update the CDK App files in your test's snapshot to
@@ -190,7 +200,7 @@ single test in a specific language. For example:
    - `tests/end-to-end/mytest/python/app.py`
    - `tests/end-to-end/mytest/typescript/app.ts`
 
-1. Review the test outputs for correctness.
+9.  Review the test outputs for correctness.
     - CDK Stacks generated by `cdk-from-cfn`
         - `tests/end-to-end/mytest/csharp/Stack.cs`
         - `tests/end-to-end/mytest/golang/stack.go`
@@ -213,10 +223,10 @@ single test in a specific language. For example:
     Note: This diff mechanism should be improved so that this manual review is
     either not necessary, or to make it easier. See [issue](TODO add link).
 
-1. To be sure your test will succeed in CI/CD, you can unset `CREATE_CFN_STACK`
+10. To be sure your test will succeed in CI/CD, you can unset `CREATE_CFN_STACK`
     and UPDATE_SNAPSHOTS env variables, and run your tests again.
 
-    ```
+    ```sh
     unset CREATE_CFN_STACK
     unset UPDATE_SNAPSHOTS
 
@@ -233,8 +243,7 @@ In these steps, assume you are updating a test called "simple".
     - The test module `end-to-end.rs`
     - A test's CloudFormation templates: `template.json` or `create-first.json`
     - A test's CDK app files: `App.ts`, `app.py`, etc.
-
-1. First, set your environment to mimic CI/CD and run the tests after your
+2. First, set your environment to mimic CI/CD and run the tests after your
    changes.
     ```sh
     unset UPDATE_SNAPSHOTS
@@ -244,17 +253,17 @@ In these steps, assume you are updating a test called "simple".
     If you changed a test, something should fail here. If you made source code
     changes, and we have good test coverage, some test should probably fail as
     well. If it doesn't, add a new test!
-1. If the input CloudFormation templates changed, make sure they are valid. Do
+3. If the input CloudFormation templates changed, make sure they are valid. Do
    step 5 from [How to add a new test](#how-to-add-a-new-test).
-1. Set `UPDATE_SNAPSHOTS=true` and run your test.
-    ```bash
-    UPDATE_SNAPSHOTS=true
+4. Set `UPDATE_SNAPSHOTS=true` and run your test.
+    ```sh
+    export UPDATE_SNAPSHOTS=true
     cargo test --test end-to-end simple
     ```
-1. If needed, iterate on failing tests until they succeed.
+5. If needed, iterate on failing tests until they succeed.
 
     For example:
-    ```bash
+    ```sh
     cargo test --test end-to-end simple::java -- --nocapture
     ```
 
@@ -262,7 +271,7 @@ In these steps, assume you are updating a test called "simple".
     as it is running. It tells `cargo` to not capture the output of the test,
     and print it to stdout right away.
 
-1. Finally, review the files that have changed using git. Files that may change:
+6. Finally, review the files that have changed using git. Files that may change:
     - `simple/typescript/Stack.ts`
         - Changes to this file indicate that `cdk-from-cfn` is now producing
           different CDK code than before. Is this expected? Did code generation
