+++ tests/end-to-end/sam_nodejs_lambda/csharp/Stack.cs
