+++ tests/end-to-end/sam_nodejs_lambda/java/src/main/java/com/myorg/Stack.java
