+++ tests/end-to-end/sam_nodejs_lambda/python/stack.py
