--- tests/end-to-end/sam_nodejs_lambda/template.yml
@@ -1,11 +1,0 @@
-Transform: AWS::Serverless-2016-10-31
-Resources:
-  MyFunction:
-    Type: AWS::Serverless::Function
-    Properties:
-      Runtime: nodejs18.x
-      Handler: index.handler
-      InlineCode: |
-        exports.handler = async (event) => {
-          console.log(event);
-        }
