+++ tests/end-to-end/sam_nodejs_lambda/typescript/stack.ts
