+++ tests/end-to-end/sam_nodejs_lambda_arr_transform/csharp/Stack.cs
