+++ tests/end-to-end/sam_nodejs_lambda_arr_transform/java/src/main/java/com/myorg/Stack.java
