+++ tests/end-to-end/sam_nodejs_lambda_arr_transform/python/stack.py
