+++ tests/end-to-end/sam_nodejs_lambda_arr_transform/template.json
@@ -1,0 +1,15 @@
+{
+  "Transform": [
+    "AWS::Serverless-2016-10-31"
+  ],
+  "Resources": {
+    "MyFunction": {
+      "Type": "AWS::Serverless::Function",
+      "Properties": {
+        "Runtime": "nodejs18.x",
+        "Handler": "index.handler",
+        "InlineCode": "exports.handler = async (event) => {\n  console.log(event);\n}\n"
+      }
+    }
+  }
+}
